usbaudio: handle sample rate overrides in proxy_prepare()
Audio playback through an external USB DAC is sometimes broken
because stream_set_new_devices() gives up when it receives a
non-zero return value from proxy_prepare(). However, EINVAL is
used in proxy_prepare() to indicate an override of requested
properties like channel counts or sample rates, according to the
following commit in system/media:

https://github.com/PixelExperience/system_media/commit/49bc03cfdc9f748c53bd989a3aaf4cd7ec4f4692

For example, in alsa_device_proxy.c, proxy->alsa_config.rate may
be configured to 192kHz, and it passes the check of
profile_is_sample_rate_valid(). When checking if alsa_config.rate
is bigger than the maximum rate, the maximum rate may be incorrectly
read as 96kHz instead of 192kHz due to a bug in the ordering of
profile->sample_rates. In this case, proxy_prepare() will try to
override alsa_config.rate to 96kHz and return EINVAL to notify the
caller. Therefore, the caller (stream_set_new_devices() in the usb
audio HAL) should probably try again with the new sample rate rather
than giving up completely because the pcm device may be perfectly
ready to be used.

alsa_config.rate differs from the max profile rate because the max
profile rate is retrived as the first item of the sample rates
array. It assumes that the array is always sorted, and the first
rate is always the maximum rate. However, that is not true in
some cases. In alsa_device_profile.c, profile sample rates are
enumerated from std_sample_rates, so it should follow the same
order as std_sample_rates. std_sample_rates put 96kHz and 88.2kHz
before 192kHz and 176.4kHz for the following reason:

"192 kHz is not first because it requires significant resources
for possibly worse quality and driver instability (depends on device)."

As a result, the profile sample rates array for a USB DAC may be
[96000, 88200, 192000, 0, ... ], where the first element is not the
max sample rate.

One solution would be to sort the std_sample_rates array from high
to low, but that might affect a whole range of other audio devices
and potentially cause regression. Instead, we can ask the usb audio
HAL to try proxy_prepare() again with the newly overriden
alsa_config.rate, if the only concern here is a change in sample
rate.

Test: boot and test USB DACs with 192kHz support

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Change-Id: I83e000c498f5e438cf75e6f651ab68aa51087e03
Signed-off-by: Joey Huab <joey@evolution-x.org>